Set repository and document abstraction ownership to core (#15)
* Initial project with dependencies

* Move interfaces to repository core

* Fix references to repository core interfaces

* Move CollectionName to repository core

* Base document keeps track of Created and Modified timestamps

* Clean up, handle null middle name

* Child and parent documents

* Datatype for measurements are floats

* Event documents

* Default collection name

* Set document properties on construction

* Document builders

* Document examples

* Dropping collection is responsibility of repository

+++ Bubbio/src/Bubbio.Core/Contracts/Events/IBiometricUpdate.cs
@@ -5,6 +5,6 @@ namespace Bubbio.Core.Contracts.Events
     public interface IBiometricUpdate : IEvent
     {
         BiometricType BiometricType { get; set; }
-        long Measurement { get; set; }
+        float Measurement { get; set; }
     }
 }
